requirements.txt
# requirements.txt

PyTorch==x.x
PyQt==x.x
numpy==x.x
matplotlib==x.x
